+++ frontend/src/pages/upload/index.js
@@ -1,8 +1,13 @@
 import React from "react";
-import UploadImageForm from "../../components/upload_new"; 
+import UploadImageForm from "../../components/upload"; 
 
 export default function testPage() {
   return (
-    <UploadImageForm />
+    <body>
+      <div>
+        <h1> THIS IS A TEST PAGE FOR UPLOADING IMAGES </h1>
+        <UploadImageForm />
+      </div>
+    </body>
   );
 };
